final refactoring to separate odoo from subgraph
