+++ codegen-subgraph.ts
@@ -3,19 +3,19 @@ import type { CodegenConfig } from "@graphql-codegen/cli";
 const config: CodegenConfig = {
   overwrite: true,
   schema: "https://api.neokingdom.org/subgraphs/name/NeokingdomDAO/vigodarzere",
-  documents: "lib/graphql/queries/subgraph/**/*.tsx",
+  documents: "lib/graphql/subgraph/queries/**/*.tsx",
   generates: {
-    "gql/": {
+    "lib/graphql/subgraph/generated/": {
       preset: "client",
       presetConfig: {
         fragmentMasking: false,
       },
     },
-    "gql/graphql.schema.json": {
+    "lib/graphql/subgraph/generated/graphql.schema.json": {
       plugins: ["introspection"],
     },
   },
-  hooks: { afterAllFileWrite: ["git add ./gql/**; npx lint-staged --allow-empty"] },
+  hooks: { afterAllFileWrite: ["git add ./lib/graphql/subgraph/generated/**; npx lint-staged --allow-empty"] },
 };
 
 export default config;
