+++ components/tokens/UserActions.tsx
@@ -3,8 +3,8 @@ import { useAccount } from "wagmi";
 
 import { CircularProgress, Divider, Grid, Paper, Typography } from "@mui/material";
 
-import { getDaoManagerQuery } from "@graphql/queries/subgraph/get-dao-manager-query";
-import { useSubgraphGraphQL } from "@graphql/subgraph";
+import { getDaoManagerQuery } from "@graphql/subgraph/queries/get-dao-manager-query";
+import { useSubgraphGraphQL } from "@graphql/subgraph/subgraph-client";
 
 import useUserBalanceAndOffers from "@hooks/useUserBalanceAndOffers";
