+++ hooks/useGetResolutions.ts
@@ -1,6 +1,6 @@
-import { getLegacyResolutionsQuery } from "@graphql/queries/subgraph/get-legacy-resolutions-query";
-import { getResolutionsQuery } from "@graphql/queries/subgraph/get-resolutions-query";
-import { useLegacySubgraphGraphQL, useSubgraphGraphQL } from "@graphql/subgraph";
+import { getLegacyResolutionsQuery } from "@graphql/subgraph/queries/get-legacy-resolutions-query";
+import { getResolutionsQuery } from "@graphql/subgraph/queries/get-resolutions-query";
+import { useLegacySubgraphGraphQL, useSubgraphGraphQL } from "@graphql/subgraph/subgraph-client";
 
 const REFRESH_EVERY_MS = 3000;
