+++ hooks/useShareholderStatus.ts
@@ -3,8 +3,8 @@ import { formatEther } from "ethers/lib/utils";
 
 import { useCallback, useMemo } from "react";
 
-import { getShareholdersInfo } from "@graphql/queries/subgraph/get-shareholders-info-query";
-import { useSubgraphGraphQL } from "@graphql/subgraph";
+import { getShareholdersInfo } from "@graphql/subgraph/queries/get-shareholders-info-query";
+import { useSubgraphGraphQL } from "@graphql/subgraph/subgraph-client";
 
 import { ShareholderStatus } from "../types";
