+++ lib/graphql/subgraph/queries/fragments/dao-manager.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const daoManagerFragment = graphql(`
   fragment daoManagerFragment on DaoManager {
