+++ lib/graphql/subgraph/queries/fragments/resolution-type.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const resolutionTypeFragment = graphql(`
   fragment resolutionTypeFragment on ResolutionType {
