+++ lib/graphql/subgraph/queries/fragments/resolution.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const resolutionFragment = graphql(`
   fragment resolutionFragment on Resolution {
