+++ lib/graphql/subgraph/queries/get-dao-manager-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getDaoManagerQuery = graphql(`
   query GetDaoManager {
