+++ lib/graphql/subgraph/queries/get-legacy-resolution-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getLegacyResolutionQuery = graphql(`
   query GetLegacyResolution($id: ID!) {
