+++ lib/graphql/subgraph/queries/get-legacy-resolutions-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getLegacyResolutionsQuery = graphql(`
   query GetLegacyResolutions {
