+++ lib/graphql/subgraph/queries/get-resolution-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getResolutionQuery = graphql(`
   query GetResolution($id: ID!) {
