+++ lib/graphql/subgraph/queries/get-resolutions-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getResolutionsQuery = graphql(`
   query GetResolutions {
