+++ lib/graphql/subgraph/queries/get-shareholders-info-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getShareholdersInfo = graphql(`
   query GetShareholdersInfo {
