+++ lib/graphql/subgraph/queries/get-subgraph-state-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getSubgraphState = graphql(`
   query GetSubgraphState {
