+++ lib/graphql/subgraph/queries/get-tokens-mintings-query.tsx
@@ -1,4 +1,4 @@
-import { graphql } from "gql";
+import { graphql } from "../generated";
 
 export const getTokenMintings = graphql(`
   query GetTokenMintings {
