+++ pages/api/pdf/resolutions/[id].ts
@@ -8,9 +8,13 @@ import React from "react";
 
 import odooClient from "@graphql/odoo";
 import { getUsersQuery } from "@graphql/queries/get-users.query";
-import { getLegacyResolutionQuery } from "@graphql/queries/subgraph/get-legacy-resolution-query";
-import { getResolutionQuery } from "@graphql/queries/subgraph/get-resolution-query";
-import { fetcherGraphqlPublic, isLegacyClientEnabled, legacyFetcherGraphqlPublic } from "@graphql/subgraph";
+import { getLegacyResolutionQuery } from "@graphql/subgraph/queries/get-legacy-resolution-query";
+import { getResolutionQuery } from "@graphql/subgraph/queries/get-resolution-query";
+import {
+  fetcherGraphqlPublic,
+  isLegacyClientEnabled,
+  legacyFetcherGraphqlPublic,
+} from "@graphql/subgraph/subgraph-client";
 
 import { getEnhancedResolutionMapper } from "@lib/resolutions/common";
 import { sessionOptions } from "@lib/session";
