+++ tsconfig.json
@@ -26,7 +26,7 @@
     "plugins": [
       {
         "name": "@0no-co/graphqlsp",
-        "schema": "./gql/graphql.schema.json"
+        "schema": "./lib/graphql/subgraph/generated/graphql.schema.json"
       }
     ]
   },
